Fix issue #696: Define API
Refactor default Array and Categorization renderers to find out
ideal structure for API
