+++ packages/default/src/additional/array/ArrayControl.tsx
@@ -1,0 +1,46 @@
+import * as React from 'react';
+import {
+  Generate,
+  composePaths,
+  DispatchRenderer
+} from '@jsonforms/core';
+
+export const ArrayControl  =
+  ({ classNames, data, label, path, resolvedSchema, onAdd }) => {
+
+    return (
+      <div className={classNames.wrapper}>
+        <fieldset className={classNames.fieldSet}>
+          <legend>
+            <button
+              className={classNames.button}
+              onClick={onAdd}
+            >
+              +
+            </button>
+            <label className={'array.label'}>
+              {label}
+            </label>
+          </legend>
+          <div className={classNames.children}>
+            {
+              data ? data.map((_child, index) => {
+
+                const generatedUi = Generate.uiSchema(resolvedSchema, 'HorizontalLayout');
+                const childPath = composePaths(path, index + '');
+
+                return (
+                  <DispatchRenderer
+                    schema={resolvedSchema}
+                    uischema={generatedUi}
+                    path={childPath}
+                    key={childPath}
+                  />
+                );
+              }) : <p>No data</p>
+            }
+          </div>
+        </fieldset>
+      </div>
+    );
+  };
