+++ packages/default/src/additional/array/index.ts
@@ -1,0 +1,10 @@
+import {
+  registerStartupRenderer,
+} from '@jsonforms/core';
+import ArrayControlRenderer from './ArrayControlRenderer';
+import { arrayTester } from './tester';
+
+export default registerStartupRenderer(
+  arrayTester,
+  ArrayControlRenderer
+);
