+++ packages/default/src/additional/categorization/index.ts
@@ -1,0 +1,8 @@
+import { registerStartupRenderer } from '@jsonforms/core';
+import CategorizationRenderer from './CategorizationRenderer';
+import { categorizationTester } from './tester';
+
+export default registerStartupRenderer(
+  categorizationTester,
+  CategorizationRenderer
+);
